Move geopmpy/plotter.py into experiments/plotting.py
