+++ integration/requirements.txt
@@ -1,0 +1,4 @@
+pillow>=9.0.0
+matplotlib>=2.2.2
+numpy>=1.14.3
+pandas>=0.23.0
