--- scripts/geopmplotter
@@ -1,12 +1,0 @@
-#!/usr/bin/env python3
-#
-#  Copyright (c) 2015 - 2022, Intel Corporation
-#  SPDX-License-Identifier: BSD-3-Clause
-#
-
-import sys
-import geopmpy.plotter
-
-if __name__ == '__main__':
-    geopmpy.plotter.main(sys.argv[1:])
-
